+++ RealmSwift-swift1.2/List.swift
@@ -56,14 +56,10 @@ public final class List<T: Object>: ListBase {
 
     // MARK: Properties
 
-    /// The Realm the objects in this list belong to, or `nil` if the list's owning
-    /// object does not belong to a realm (the list is standalone).
+    /// The Realm the objects in this list belong to, or `nil` if the list's
+    /// owning object does not belong to a realm (the list is standalone).
     public var realm: Realm? {
-        if let rlmRealm = _rlmArray.realm {
-            return Realm(rlmRealm)
-        } else {
-            return nil
-        }
+        return map(_rlmArray.realm) { Realm($0) }
     }
 
     /// Indicates if the list can no longer be accessed.